Move CqlQuery class to http client package CIRC-308
