+++ src/main/java/org/folio/circulation/domain/ConfigurationRepository.java
@@ -1,7 +1,7 @@
 package org.folio.circulation.domain;
 
 import static org.folio.circulation.domain.MultipleRecords.from;
-import static org.folio.circulation.support.CqlQuery.exactMatch;
+import static org.folio.circulation.support.http.client.CqlQuery.exactMatch;
 
 import io.vertx.core.json.JsonObject;
 import java.util.Collection;
@@ -11,7 +11,7 @@ import java.util.function.Function;
 import org.folio.circulation.domain.anonymization.config.LoanAnonymizationConfiguration;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.CollectionResourceClient;
-import org.folio.circulation.support.CqlQuery;
+import org.folio.circulation.support.http.client.CqlQuery;
 import org.folio.circulation.support.Result;
 import org.joda.time.DateTimeZone;
