+++ src/main/java/org/folio/circulation/domain/notice/schedule/ScheduledNoticesRepository.java
@@ -2,8 +2,8 @@ package org.folio.circulation.domain.notice.schedule;
 
 import static java.util.function.Function.identity;
 import static org.folio.circulation.domain.notice.schedule.JsonScheduledNoticeMapper.mapToJson;
-import static org.folio.circulation.support.CqlQuery.exactMatch;
-import static org.folio.circulation.support.CqlQuery.exactMatchAny;
+import static org.folio.circulation.support.http.client.CqlQuery.exactMatch;
+import static org.folio.circulation.support.http.client.CqlQuery.exactMatchAny;
 import static org.folio.circulation.support.http.CommonResponseInterpreters.noContentRecordInterpreter;
 import static org.folio.circulation.support.http.ResponseMapping.flatMapUsingJson;
 import static org.folio.circulation.support.http.ResponseMapping.forwardOnFailure;
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 import org.folio.circulation.domain.MultipleRecords;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.CollectionResourceClient;
-import org.folio.circulation.support.CqlQuery;
+import org.folio.circulation.support.http.client.CqlQuery;
 import org.folio.circulation.support.CqlSortBy;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.http.client.Response;
