+++ src/main/java/org/folio/circulation/support/http/client/CqlQuery.java
@@ -1,4 +1,4 @@
-package org.folio.circulation.support;
+package org.folio.circulation.support.http.client;
 
 import static java.lang.String.format;
 import static java.lang.String.join;
@@ -16,6 +16,8 @@ import java.util.List;
 import java.util.Objects;
 
 import org.apache.commons.lang3.StringUtils;
+import org.folio.circulation.support.CqlSortBy;
+import org.folio.circulation.support.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -80,7 +82,7 @@ public class CqlQuery {
     return new CqlQuery(query, sortBy);
   }
 
-  Result<String> encode() {
+  public Result<String> encode() {
     final String sortedQuery = sortBy.applyTo(query);
 
     log.info("Encoding query {}", sortedQuery);
